+++ idp_engine/Theory.py
@@ -197,10 +197,10 @@ class Theory(object):
                 instantiables = defin.get_instantiables(for_explain=True)
                 defin.add_def_constraints(instantiables, self, def_constraints)
 
-            todo = chain(self.constraints, chain(*def_constraints.values()))
-            for constraint in todo:
+            for constraint in chain([c.interpret(self) for c in self.constraints],
+                                    chain(*def_constraints.values())):
                 p = constraint.reified(self)
-                self.expl_reifs[p] = (constraint.original.interpret(self).translate(self), constraint)
+                self.expl_reifs[p] = (constraint.translate(self), constraint)
                 self._reif.add(Implies(p, self.expl_reifs[p][0]))
 
         return self._reif
@@ -777,8 +777,7 @@ class Theory(object):
 
         result = solver.check([z3_form for z3_form, (_, expr) in ps.items() if
                                     not (expr and expr.code in self.ignored_laws)])
-        assert result == unsat, ("Incorrect solver result during explain inference. "
-                                 "This may be due to floating point imprecision.")
+        assert result == unsat, ("Theory is satisfiable: nothing to explain.")
         unsatcore = solver.unsat_core()
 
         solver.pop()